Move most cut-specific stuff to RetailSubmap
