+++ src/main/java/legend/game/Scus94491BpeSegment_8005.java
@@ -41,8 +41,6 @@ public final class Scus94491BpeSegment_8005 {
     712, 712, 712, 712, 712, 712, 712, 712,
   };
 
-  public static int _80050274 = -1;
-
   public static final UvAdjustmentMetrics14[] _8005027c = {
     UvAdjustmentMetrics14.NONE,
     new UvAdjustmentMetrics14( 1, 320, 496, 320, 256),
