+++ src/main/java/legend/game/submap/MapTransitionData4c.java
@@ -1,7 +1,7 @@
 package legend.game.submap;
 
 import java.util.Arrays;
-import java.util.function.Consumer;
+import java.util.function.BiConsumer;
 
 public class MapTransitionData4c {
   /** The primitives bordering exits to the world map */
@@ -9,12 +9,16 @@ public class MapTransitionData4c {
   public int worldMapExitCount_40;
   /** Normally true, latched to false once the player steps onto a boundary that leads to the world map. Once false, stops updating the collided index while Dart leaves the map. */
   public boolean shouldUpdateCollidedWith_44;
-  public Consumer<MapTransitionData4c> collidedPrimitiveSetterCallback_48;
+  public BiConsumer<SubmapObject210, MapTransitionData4c> collidedPrimitiveSetterCallback_48;
+
+  public void run(final SubmapObject210 sobj) {
+    this.collidedPrimitiveSetterCallback_48.accept(sobj, this);
+  }
 
   public void clear() {
     Arrays.fill(this.worldMapExitPrimitiveIndices_00, 0);
     this.worldMapExitCount_40 = 0;
     this.shouldUpdateCollidedWith_44 = false;
-    this.collidedPrimitiveSetterCallback_48 = null;
+    this.collidedPrimitiveSetterCallback_48 = (sobj, transitionData) -> { };
   }
 }
