+++ src/main/java/legend/game/submap/SubmapMusic08.java
@@ -1,12 +1,12 @@
 package legend.game.submap;
 
 public class SubmapMusic08 {
-  public int submapCut_00;
+  public int submapId_00;
   public int musicIndex_02;
   public int[] submapCuts_04;
 
-  public SubmapMusic08(final int submapCut, final int musicIndex, final int... submapCuts) {
-    this.submapCut_00 = submapCut;
+  public SubmapMusic08(final int submapId, final int musicIndex, final int... submapCuts) {
+    this.submapId_00 = submapId;
     this.musicIndex_02 = musicIndex;
     this.submapCuts_04 = submapCuts;
   }
